Pankesh v2

+++ Library_Management/library_management/Library/views.py
@@ -1,53 +1,34 @@
 from asyncio.windows_events import NULL
-
+import json
+from msilib.schema import Error
+import re
+from django.contrib import messages
 from telnetlib import STATUS
 from urllib import response
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from .models import Book,MyBook
 from django.contrib.auth.decorators import login_required
-
-# books = [   {
-#         'BookTitle':'Pride and Prejudice',
-#         'Author':'Jane Austen, Anna Quindlen',
-#         'Publisher':'T. Egerton, Whitehall',
-#         'Genre':'Romance',
-#         'Year':'1813',
-#         'Language':'English',
-#         'Description':'The Pride of Jane Austen! The story is set in England in the late 18th century. Charles Bingley, a wealthy and charismatic single man, moves to the Netherfield estate, and when he does, the residents there are very thrilled',
-#         'Tags':['Romance','Fiction'],
-#         'ISBN':'9780140430721'
-#     },
-#     {
-#         'BookTitle':'In Search of Lost Time',
-#         'Author':'Marcel Proust',
-#         'Publisher':'Grasset and Gallimard',
-#         'Genre':'Modernist',
-#         'Year':'1913',
-#         'Language':'English',
-#         'Description':'In Search of Lost Time follows the narrators recollections of childhood and experiences into adulthood in the late 19th-century and early 20th-century high-society France',
-#         'Tags':['French','Historic'],
-#         'ISBN':'9780099362210'
-#     },
-#     {
-#         'BookTitle':'To Kill a Mockingbird',
-#         'BookTitle':'To Kill a Mockingbird',
-#         'Author':'Harper Lee',
-#         'Publisher':'J. B. Lippincott & Co.',
-#         'Genre':'Southern Gothic',
-#         'Year':'1960',
-#         'Language':'English',
-#         'Description':'To Kill a Mockingbird takes place in the fictional town of Maycomb, Alabama, during the Great Depression.',
-#         'Tags':['Bildungsroman','Legal'],
-#         'ISBN':'9782253115847'
-#     }
-# ]
+from django.core.paginator import Paginator
 
 
 @login_required
 def booklist(request):
+    if request.method == 'POST':
+        search_key=request.POST.get('search',None)
+        if not search_key:
+            books=Book.objects.all()
+        else:
+            books=Book.objects.filter(BookTitle__contains=search_key)
+    else:
+        books=Book.objects.all()
+
+
+    paginator = Paginator(books, 9)
+    page_number = request.GET.get('page',1)
+    page_obj = paginator.get_page(page_number) 
     context ={
-        'books':Book.objects.all()
+        'page_obj': page_obj
     }
     return render(request, 'Library/booklist.html', context)
 
@@ -56,22 +37,36 @@ def booklist(request):
 
 @login_required
 def IssueBook(request):
-    book = request.POST.get("id",None)
-    bk=Book.objects.get(pk=book)
-    bk.IssuedBy = request.user
-    bk.save()
-    print("book saved")
 
 
-    oldbooks= MyBook.objects.filter(book=bk)
-    print(oldbooks)
-    oldbooks.delete()
+    if  MyBook.objects.filter(user=request.user).count()>=1:
+        
+        response = HttpResponse(json.dumps({'msg': f'You cant issue more books. Please return previous books'}), 
+        content_type='application/json')
+        response.status_code=400
+        return response
+        
+    else:
 
-    mybk= MyBook(book=bk,user=request.user)
-    mybk.save()
-    print("My book saved")
+        book = request.POST.get("id",None)
+        bk=Book.objects.get(pk=book)
+        bk.IssuedBy = request.user
+        bk.save()
+        print("book saved")
 
-    return HttpResponse(status=200)
+
+        oldbooks= MyBook.objects.filter(book=bk)
+        print(oldbooks)
+        oldbooks.delete()
+
+        mybk= MyBook(book=bk,user=request.user)
+        mybk.save()
+        print("My book saved")
+
+        response = HttpResponse(json.dumps({'msg': f'Your book has been issued'}), 
+        content_type='application/json')
+        response.status_code=200    
+        return response
 
 
 
@@ -79,7 +74,7 @@ def IssueBook(request):
 @login_required
 def mybooks(request):
     context ={
-        'mybooks':MyBook.objects.all()
+        'mybooks':MyBook.objects.filter(user=request.user)
     }
     return render(request, 'Library/mybooks.html', context)
